Got the core pieces working.
got a grid that holds all the divs for coloring. script builds the div
and have a basic color wheel with a display to show current color. a
button to clear the box and a button to remove the grid lines or add
them back. stay tuned for styling and addition changes.
TODO: Add a option to make custom canvas size and background color.
TODO: style the site to looks prettier.